requirements.txt
pytest==7.1.2
Appium-Python-Client==2.2.0
selenium~=4.2.0